[protocol3] Increased max number of tokens to 1024

+++ packages/loopring_v3/contracts/impl/libexchange/ExchangeDeposits.sol
@@ -111,7 +111,7 @@ library ExchangeDeposits
                                       // In the circuit we always overwrite the public keys in
                                       // the Account leaf with the data given onchain.
                     account.pubKeyY,
-                    uint8(tokenID),
+                    uint16(tokenID),
                     amount
                 )
             ),
